Refactor heater inventory helpers
