+++ tests/test_installation_snapshot.py
@@ -5,9 +5,9 @@ from typing import Any
 
 import pytest
 
-from custom_components.termoweb.heater_inventory import build_heater_inventory_details
 from custom_components.termoweb.installation import InstallationSnapshot, ensure_snapshot
 from custom_components.termoweb.inventory import (
+    build_heater_inventory_details,
     build_node_inventory,
     heater_sample_subscription_targets,
 )
